add basic Project API test
