+++ EnviroMonitorWeb/api/tests/test_views.py
@@ -1,18 +1,17 @@
-from factory.django import DjangoModelFactory
+from rest_framework.reverse import reverse
+from rest_framework.test import APIRequestFactory, APITestCase
+from rest_framework.status import HTTP_201_CREATED, HTTP_200_OK
 
-from api.models import Station, Metering, Project
+from .factories import ProjectFactory
+from api.serializers import ProjectSerializer
 
-
-class ProjectFactory(DjangoModelFactory):
-    class Meta:
-        model = Project
-
-
-class StationFactory(DjangoModelFactory):
-    class Meta:
-        model = Station
+import pytest
 
 
-class MeteringFactory(DjangoModelFactory):
-    class Meta:
-        model = Metering
+class ProjectApiTests(APITestCase):
+
+    @pytest.mark.django_db
+    def test_project_list(self):
+        api_url = reverse('project-list')
+        api_response = self.client.get(api_url)
+        self.assertEqual(api_response.status_code, HTTP_200_OK)
